首页重定向修改

+++ php-code/application/hooks/loginFilter.php
@@ -17,16 +17,18 @@ class LoginFilter
     
     public function checkLogin()
     {
-        
+        //首页默认重定向到mobile/home 
+        $this->CI = & get_instance();
         $uri = $_SERVER['REQUEST_URI'];
+        if ($uri == "/") {
+            $this->CI->load->helper('url');
+            redirect('mobile/home','refresh');
+        }
         foreach ($this -> noNeedLogin as $value) {
             if (strstr($uri,$value)){
                 return;
             }
         }
-        
-        $this->CI = & get_instance();
-        $class = $this->CI -> router->class;
         
         $this->CI->load->library('session');
         $this->CI->load->model('sys/sys_token_model');
@@ -37,15 +39,19 @@ class LoginFilter
         /*
          * 提供两种验证方式：1、请求头token,2、session
          */
-        $token = $_SERVER['HTTP_TOKEN'];
+        if (array_key_exists('HTTP_TOKEN', $_SERVER)) {
+            $token = $_SERVER['HTTP_TOKEN'];
+            $token_database = $this->CI->sys_token_model->findByToken($token);
+            $preTime = $token_database['gmt_modified'];
+        }
         $user = $this->CI->session->tempdata('user');
         
-        $token_database = $this->CI->sys_token_model->findByToken($token);
         $time = date('YmdHis');
-        $preTime = $token_database['gmt_modified'];
 
         if (empty($token_database)) {
-            $this->CI->sys_token_model->deleteByToken($token);
+            if (!empty($token)) {
+                $this->CI->sys_token_model->deleteByToken($token);
+            }
             echo json_encode(array(
                 'code' => '101',
                 'msg' => '用户没有登录',
